# v1.1.1
- The old search menu can be opened with a click on a green arrow on the bottom right of the search popup
- Android users can finally input the correct amount of characters into text fields
- Old tier label displays on top of the epic/legendary/mythic fire
- The tier button/label is given a texture/color and content that matches the tier in the local cache, making it feel like the rating is instantly loaded

+++ src/GDDLRating.cpp
@@ -8,5 +8,5 @@ GDDLRating::GDDLRating(json levelData) {
     this->enjoymentCount = levelData["EnjoymentCount"];
     this->submissionCount = levelData["SubmissionCount"];
 
-    this->roundedRating = round(this->rating);
+    this->roundedRating = static_cast<int>(round(this->rating));
 }
